Move LedgerState into the block module
