+++ zebra-chain/src/lib.rs
@@ -34,42 +34,5 @@ pub mod transaction;
 pub mod transparent;
 pub mod work;
 
-#[derive(Debug, Clone, Copy)]
 #[cfg(any(test, feature = "proptest-impl"))]
-#[non_exhaustive]
-/// The configuration data for proptest when generating arbitrary chains
-pub struct LedgerState {
-    /// The tip height of the block or start of the chain
-    pub tip_height: block::Height,
-    is_coinbase: bool,
-    /// The network to generate fake blocks for
-    pub network: parameters::Network,
-}
-
-#[cfg(any(test, feature = "proptest-impl"))]
-impl LedgerState {
-    /// Construct a new ledger state for generating arbitrary chains via proptest
-    pub fn new(tip_height: block::Height, network: parameters::Network) -> Self {
-        Self {
-            tip_height,
-            is_coinbase: true,
-            network,
-        }
-    }
-}
-
-#[cfg(any(test, feature = "proptest-impl"))]
-impl Default for LedgerState {
-    fn default() -> Self {
-        let network = parameters::Network::Mainnet;
-        let tip_height = parameters::NetworkUpgrade::Canopy
-            .activation_height(network)
-            .unwrap();
-
-        Self {
-            tip_height,
-            is_coinbase: true,
-            network,
-        }
-    }
-}
+pub use block::LedgerState;
